Added floor tiles with less detail.
The previous tiles used for the floor of every map where very detailed. That made them hard on the eyes when the player was moving around. New, less detailed versions of those tiles were added and now make up most of the floor of each arena. The old tiles are still around, decorating certain sections.